chore(build-deploy.yml): add GitHub Actions workflow to build and deploy the Vue branch to GitHub Pages
A new GitHub Actions workflow file, `build-deploy.yml`, has been added. This workflow is triggered on push events to the `vue` branch. It consists of a single job, `build-and-deploy`, which runs on the latest version of Ubuntu. The steps in the job include:
- Checking out the repository
- Setting up Node.js version 14
- Caching the npm dependencies
- Building the project by running `npm install` and `npm run build`
- Copying the `index.html` file to `404.html` in the `dist` folder
- Deploying the `dist` folder to the `gh-pages` branch using the `github-pages-deploy-action` action

This workflow enables automatic building and deployment of the Vue branch to GitHub Pages whenever changes are pushed.